Function arrow_ord::cmp::apply_op

fn apply_op<T: ArrayOrd>(
    l: T,
    l_s: Option<usize>,
    r: T,
    r_s: Option<usize>,
    neg: bool,
    op: impl Fn(T::Item, T::Item) -> bool,
) -> BooleanBuffer

Applies op to possibly scalar ArrayOrd

If l is scalar l_s will be Some(idx) where idx is the index of the scalar value in l If r is scalar r_s will be Some(idx) where idx is the index of the scalar value in r

If neg is true the result of op will be negated