Module arrow_ord::comparison

Comparison kernels for Arrays.

These kernels can leverage SIMD if available on your system. Currently no runtime detection is provided, you should enable the specific SIMD intrinsics using RUSTFLAGS="-C target-feature=+avx2" for example. See the documentation here for more information.

Macros

dyn_compare_scalar 🔒

dyn_compare_utf8_scalar 🔒

try_to_type 🔒

Calls $RIGHT.$TY() (e.g. right.to_i128()) with a nice error message. Type of expression is Result<.., ArrowError>

typed_cmp_dict_non_dict 🔒

typed_compares 🔒

typed_dict_cmp 🔒

Applies $OP to $LEFT and $RIGHT which are two dictionaries which have (the same) key type $KT

typed_dict_compares 🔒

typed_dict_non_dict_cmp 🔒

typed_dict_string_array_cmp 🔒

Functions

cmp_dict

Perform given operation on two DictionaryArrays which value type is primitive type. Returns an error if the two arrays have different value type

cmp_dict_binary

Perform the given operation on two DictionaryArrays which value type is DataType::Binary or DataType::LargeBinary.

cmp_dict_binary_array 🔒

Perform given operation on DictionaryArray and GenericBinaryArray. The value type of DictionaryArray is same as GenericBinaryArray’s type.

cmp_dict_bool

Perform the given operation on two DictionaryArrays which value type is DataType::Boolean.

cmp_dict_boolean_array 🔒

Perform given operation on DictionaryArray and BooleanArray. The value type of DictionaryArray is same as BooleanArray’s type.

cmp_dict_primitive 🔒

Perform given operation on DictionaryArray and PrimitiveArray. The value type of DictionaryArray is same as PrimitiveArray’s type.

cmp_dict_string_array 🔒

Perform given operation on DictionaryArray and GenericStringArray. The value type of DictionaryArray is same as GenericStringArray’s type.

cmp_dict_utf8

Perform the given operation on two DictionaryArrays which value type is DataType::Utf8 or DataType::LargeUtf8.

cmp_primitive_array 🔒

compare_op 🔒

Helper function to perform boolean lambda function on values from two array accessors, this version does not attempt to use SIMD.

compare_op_scalar 🔒

Helper function to perform boolean lambda function on values from array accessor, this version does not attempt to use SIMD.

eq

Perform left == right operation on two [PrimitiveArray]s.

eq_binary

Perform left == right operation on [BinaryArray] / [LargeBinaryArray].

eq_binary_scalar

Perform left == right operation on [BinaryArray] / [LargeBinaryArray] and a scalar

eq_bool

Perform left == right operation on [BooleanArray]

eq_bool_scalar

Perform left == right operation on [BooleanArray] and a scalar

eq_dyn

Perform left == right operation on two (dynamic) [Array]s.

eq_dyn_binary_scalar

Perform left == right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

eq_dyn_bool_scalar

Perform left == right operation on an array and a numeric scalar value.

eq_dyn_scalar

Perform left == right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

eq_dyn_utf8_scalar

Perform left == right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

eq_scalar

Perform left == right operation on a [PrimitiveArray] and a scalar value.

eq_utf8

Perform left == right operation on [StringArray] / [LargeStringArray].

eq_utf8_scalar

Perform left == right operation on [StringArray] / [LargeStringArray] and a scalar.

gt

Perform left > right operation on two [PrimitiveArray]s. Non-null values are greater than null values.

gt_binary

Perform left > right operation on [BinaryArray] / [LargeBinaryArray].

gt_binary_scalar

Perform left > right operation on [BinaryArray] / [LargeBinaryArray] and a scalar.

gt_bool

Perform left > right operation on [BooleanArray]

gt_bool_scalar

Perform left > right operation on [BooleanArray] and a scalar

gt_dyn

Perform left > right operation on two (dynamic) [Array]s.

gt_dyn_binary_scalar

Perform left > right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

gt_dyn_bool_scalar

Perform left > right operation on an array and a numeric scalar value. Supports BooleanArrays.

gt_dyn_scalar

Perform left > right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

gt_dyn_utf8_scalar

Perform left > right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

gt_eq

Perform left >= right operation on two [PrimitiveArray]s. Non-null values are greater than null values.

gt_eq_binary

Perform left >= right operation on [BinaryArray] / [LargeBinaryArray].

gt_eq_binary_scalar

Perform left >= right operation on [BinaryArray] / [LargeBinaryArray] and a scalar.

gt_eq_bool

Perform left >= right operation on [BooleanArray]

gt_eq_bool_scalar

Perform left >= right operation on [BooleanArray] and a scalar

gt_eq_dyn

Perform left >= right operation on two (dynamic) [Array]s.

gt_eq_dyn_binary_scalar

Perform left >= right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

gt_eq_dyn_bool_scalar

Perform left >= right operation on an array and a numeric scalar value. Supports BooleanArrays.

gt_eq_dyn_scalar

Perform left >= right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

gt_eq_dyn_utf8_scalar

Perform left >= right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

gt_eq_scalar

Perform left >= right operation on a [PrimitiveArray] and a scalar value. Non-null values are greater than null values.

gt_eq_utf8

Perform left >= right operation on [StringArray] / [LargeStringArray].

gt_eq_utf8_scalar

Perform left >= right operation on [StringArray] / [LargeStringArray] and a scalar.

gt_scalar

Perform left > right operation on a [PrimitiveArray] and a scalar value. Non-null values are greater than null values.

gt_utf8

Perform left > right operation on [StringArray] / [LargeStringArray].

gt_utf8_scalar

Perform left > right operation on [StringArray] / [LargeStringArray] and a scalar.

in_list

Checks if a [GenericListArray] contains a value in the [PrimitiveArray]

in_list_utf8

Checks if a [GenericListArray] contains a value in the [GenericStringArray]

lt

Perform left < right operation on two [PrimitiveArray]s. Null values are less than non-null values.

lt_binary

Perform left < right operation on [BinaryArray] / [LargeBinaryArray].

lt_binary_scalar

Perform left < right operation on [BinaryArray] / [LargeBinaryArray] and a scalar.

lt_bool

Perform left < right operation on [BooleanArray]

lt_bool_scalar

Perform left < right operation on [BooleanArray] and a scalar

lt_dyn

Perform left < right operation on two (dynamic) [Array]s.

lt_dyn_binary_scalar

Perform left < right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

lt_dyn_bool_scalar

Perform left < right operation on an array and a numeric scalar value. Supports BooleanArrays.

lt_dyn_scalar

Perform left < right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

lt_dyn_utf8_scalar

Perform left < right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

lt_eq

Perform left <= right operation on two [PrimitiveArray]s. Null values are less than non-null values.

lt_eq_binary

Perform left <= right operation on [BinaryArray] / [LargeBinaryArray].

lt_eq_binary_scalar

Perform left <= right operation on [BinaryArray] / [LargeBinaryArray] and a scalar.

lt_eq_bool

Perform left <= right operation on [BooleanArray]

lt_eq_bool_scalar

Perform left <= right operation on [BooleanArray] and a scalar

lt_eq_dyn

Perform left <= right operation on two (dynamic) [Array]s.

lt_eq_dyn_binary_scalar

Perform left <= right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

lt_eq_dyn_bool_scalar

Perform left <= right operation on an array and a numeric scalar value. Supports BooleanArrays.

lt_eq_dyn_scalar

Perform left <= right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

lt_eq_dyn_utf8_scalar

Perform left <= right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

lt_eq_scalar

Perform left <= right operation on a [PrimitiveArray] and a scalar value. Null values are less than non-null values.

lt_eq_utf8

Perform left <= right operation on [StringArray] / [LargeStringArray].

lt_eq_utf8_scalar

Perform left <= right operation on [StringArray] / [LargeStringArray] and a scalar.

lt_scalar

Perform left < right operation on a [PrimitiveArray] and a scalar value. Null values are less than non-null values.

lt_utf8

Perform left < right operation on [StringArray] / [LargeStringArray].

lt_utf8_scalar

Perform left < right operation on [StringArray] / [LargeStringArray] and a scalar.

neq

Perform left != right operation on two [PrimitiveArray]s.

neq_binary

Perform left != right operation on [BinaryArray] / [LargeBinaryArray].

neq_binary_scalar

Perform left != right operation on [BinaryArray] / [LargeBinaryArray] and a scalar.

neq_bool

Perform left != right operation on [BooleanArray]

neq_bool_scalar

Perform left != right operation on [BooleanArray] and a scalar

neq_dyn

Perform left != right operation on two (dynamic) [Array]s.

neq_dyn_binary_scalar

Perform left != right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

neq_dyn_bool_scalar

Perform left != right operation on an array and a numeric scalar value. Supports BooleanArrays.

neq_dyn_scalar

Perform left != right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

neq_dyn_utf8_scalar

Perform left != right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

neq_scalar

Perform left != right operation on a [PrimitiveArray] and a scalar value.

neq_utf8

Perform left != right operation on [StringArray] / [LargeStringArray].

neq_utf8_scalar

Perform left != right operation on [StringArray] / [LargeStringArray] and a scalar.

no_simd_compare_op

Evaluate op(left, right) for [PrimitiveArray]s using a specified comparison function.

no_simd_compare_op_scalar

Evaluate op(left, right) for [PrimitiveArray] and scalar using a specified comparison function.

simd_compare_op 🔒

Helper function to perform boolean lambda function on values from two arrays using SIMD.

simd_compare_op_scalar 🔒

Helper function to perform boolean lambda function on values from an array and a scalar value using SIMD.

try_to_type_result 🔒

unary_cmp

Applies an unary and infallible comparison function to a primitive array.

unpack_dict_comparison 🔒

unpacks the results of comparing left.values (as a boolean)

utf8_empty 🔒