arrow_ord

Module cmp

Comparison kernels for Arrays.

These kernels can leverage SIMD if available on your system. Currently no runtime detection is provided, you should enable the specific SIMD intrinsics using RUSTFLAGS="-C target-feature=+avx2" for example. See the documentation here for more information.

Enums

Op 🔒

Traits

ArrayOrd 🔒

Functions

apply 🔒

Perform a potentially vectored op on the provided ArrayOrd

apply_op 🔒

Applies op to possibly scalar ArrayOrd

apply_op_vectored 🔒

Applies op to possibly scalar ArrayOrd with the given indices

collect_bool 🔒

Invokes f with values 0..len collecting the boolean results into a new BooleanBuffer

compare_byte_view

Compares two [GenericByteViewArray] at index left_idx and right_idx

compare_byte_view_unchecked⚠Deprecated

Comparing two [GenericByteViewArray] at index left_idx and right_idx

compare_op 🔒

Perform op on the provided Datum

distinct

Perform left IS DISTINCT FROM right operation on two [Datum]

eq

Perform left == right operation on two [Datum].

gt

Perform left > right operation on two [Datum].

gt_eq

Perform left >= right operation on two [Datum].

lt

Perform left < right operation on two [Datum].

lt_eq

Perform left <= right operation on two [Datum].

neq

Perform left != right operation on two [Datum].

not_distinct

Perform left IS NOT DISTINCT FROM right operation on two [Datum]

take_bits 🔒

Perform a take operation on buffer with the given dictionary