arrow_ord

Module ord

Contains functions and function factories to compare arrays.

Functions

child_opts ğŸ”’

If parent sort order is descending we need to invert the value of nulls_first so that when the parent is sorted based on the produced ranks, nulls are still ordered correctly

compare ğŸ”’

compare_boolean ğŸ”’

compare_byte_view ğŸ”’

compare_bytes ğŸ”’

compare_dict ğŸ”’

compare_fixed_list ğŸ”’

compare_impl ğŸ”’

compare_list ğŸ”’

compare_list_view ğŸ”’

compare_map ğŸ”’

compare_primitive ğŸ”’

compare_run_end_encoded ğŸ”’

compare_struct ğŸ”’

compare_union ğŸ”’

make_comparator

Returns a comparison function that compares two values at two different positions between the two arrays.

Type Aliases

DynComparator

Compare the values at two arbitrary indices in two arrays.