arrow_ord

Module rank

Provides rank function to assign a rank to each value in an array

Functions

boolean_rank ğŸ”’

bytes_rank ğŸ”’

can_rank ğŸ”’

Whether arrow_ord::rank can rank an array of given data type.

get_boolean_rank_index ğŸ”’

Return the index for the rank when ranking boolean array

primitive_rank ğŸ”’

rank

Assigns a rank to each value in array based on its position in the sorted order

rank_impl ğŸ”’