arrow_ord
Module sort

Defines sort kernel for ArrayRef

Structs

LexicographicalComparator

A lexicographical comparator that wraps given array data (columns) and can lexicographically compare data at given two indices. The lifetime is the same at the data wrapped.

SortColumn

One column to be used in lexicographical sort

SortOptions

Options that define the sort order of a given column

Functions

can_rank 🔒

Whether arrow_ord::rank can rank an array of given data type.

can_sort_to_indices 🔒

Whether sort_to_indices can sort an array of given data type.

child_rank 🔒

Computes the rank for a set of child values

lexsort

Sort a list of ArrayRef using SortOptions provided for each array.

lexsort_to_indices

Sort elements lexicographically from a list of ArrayRef into an unsigned integer (UInt32Array) of indices.

partial_sort

It’s unstable_sort, may not preserve the order of equal elements

partition_validity 🔒

sort

Sort the ArrayRef using SortOptions.

sort_boolean 🔒

sort_byte_view 🔒

sort_bytes 🔒

sort_dictionary 🔒

sort_fixed_size_binary 🔒

sort_fixed_size_list 🔒

sort_impl 🔒

sort_limit

Sort the ArrayRef partially.

sort_list 🔒

sort_native_type 🔒

sort_primitive 🔒

sort_run 🔒

sort_run_downcasted 🔒

sort_run_inner 🔒

sort_run_to_indices 🔒

sort_to_indices

Sort elements from ArrayRef into an unsigned integer (UInt32Array) of indices. Floats are sorted using IEEE 754 totalOrder. limit is an option for partial_sort.

sort_unstable_by 🔒

we can only do this if the T is primitive