Module arrow_ord::sort

Defines sort kernel for ArrayRef

Structs

LexicographicalComparator

A lexicographical comparator that wraps given array data (columns) and can lexicographically compare data at given two indices. The lifetime is the same at the data wrapped.

SortColumn

One column to be used in lexicographical sort

SortOptions

Options that define the sort order of a given column

Functions

cmp 🔒

cmp_array 🔒

Compare two Arrays based on the ordering defined in build_compare

insert_valid_values 🔒

lexsort

Sort a list of ArrayRef using SortOptions provided for each array.

lexsort_to_indices

Sort elements lexicographically from a list of ArrayRef into an unsigned integer (UInt32Array) of indices.

partial_sort

It’s unstable_sort, may not preserve the order of equal elements

partition_validity 🔒

sort

Sort the ArrayRef using SortOptions.

sort_binary 🔒

sort_boolean 🔒

Sort boolean values

sort_limit

Sort the ArrayRef partially.

sort_list 🔒

sort_list_inner 🔒

sort_native_type 🔒

sort_primitive 🔒

Sort primitive values

sort_primitive_dictionary 🔒

Sort dictionary encoded primitive values

sort_primitive_inner 🔒

sort_run 🔒

sort_run_downcasted 🔒

sort_run_inner 🔒

sort_run_to_indices 🔒

sort_string 🔒

Sort strings

sort_string_dictionary 🔒

Sort dictionary encoded strings

sort_string_helper 🔒

shared implementation between dictionary encoded and plain string arrays

sort_to_indices

Sort elements from ArrayRef into an unsigned integer (UInt32Array) of indices. For floating point arrays any NaN values are considered to be greater than any other non-null value. limit is an option for partial_sort.

sort_unstable_by 🔒

we can only do this if the T is primitive

sort_valids 🔒

sort_valids_array 🔒

sorted_rank 🔒

Given a list of indices that yield a sorted order, returns the ordered rank of each index

Type Definitions

LexicographicalCompareItem 🔒